test: add CP test and fix typecast issue

+++ src/instructions.cpp
@@ -300,7 +300,7 @@ void CPU::AND_A_n8(uint8_t n8) {
 // CP A, r8
 void CPU::CP_A_r8(R8_PTR r) {
   uint8_t reg_value = read_reg(r);
-  uint8_t result = reg.a - reg_value;
+  uint16_t result = (uint8_t)(reg.a - reg_value);
   set_z_flag(result);
   flag_value(N, 1);
   flag_value(H, ((reg.a & 0x0f) - (reg_value & 0x0f)) < 0);
@@ -310,7 +310,7 @@ void CPU::CP_A_r8(R8_PTR r) {
 // CP A, [HL]
 void CPU::CP_A_HL() {
   uint8_t reg_value = read_byte(reg.hl);
-  uint8_t result = reg.a - reg_value;
+  uint16_t result = (uint8_t)(reg.a - reg_value);
   set_z_flag(result);
   flag_value(N, 1);
   flag_value(H, ((reg.a & 0x0f) - (reg_value & 0x0f)) < 0);
@@ -319,7 +319,7 @@ void CPU::CP_A_HL() {
 
 // CP A, n8
 void CPU::CP_A_n8(uint8_t n8) {
-  uint8_t result = reg.a - n8;
+  uint16_t result = (uint8_t)(reg.a - n8);
   set_z_flag(result);
   flag_value(N, 1);
   flag_value(H, ((reg.a & 0x0f) - (n8 & 0x0f)) < 0);
